backend_fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "FastAPI adapter for hexagonal architecture"
requires-python = ">=3.13"
dependencies = [
    "core",
    "infrastructure",
    "fastapi",
    "python-dotenv>=1.0.0",
]
